requirements.txt
feedparser==6.0.10
yt-dlp==2026.03.03
requests==2.32.3
beautifulsoup4==4.12.3
tiktoken==0.7.0
openai==1.61.0
edge-tts==6.1.9
jinja2==3.1.6
python-dotenv==1.0.1
aiohttp==3.11.13
